refactor(tasks): move Go1 joystick owner [skip ci]
