+++ src/unilab/tasks/locomotion/go1/__init__.py
@@ -1,3 +1,9 @@
+from .joystick import Go1JoystickCfg, Go1WalkTask
 from .rough import Go1JoystickRoughCfg, Go1JoystickRoughEnv
 
-__all__ = ["Go1JoystickRoughCfg", "Go1JoystickRoughEnv"]
+__all__ = [
+    "Go1JoystickCfg",
+    "Go1JoystickRoughCfg",
+    "Go1JoystickRoughEnv",
+    "Go1WalkTask",
+]
